firmware/active_sta_py/requirements.txt
scapy>=2.5.0
python-dotenv>=0.19.0
pywifi>=1.1.12
numpy>=1.21.0
argparse>=1.4.0
